+++ InGame/Entities/Enemies/IEnemy.cs
@@ -8,6 +8,6 @@ public interface IEnemy : IEntity, IDamageable, IDieable, IRemovable
     /// <summary>
     /// この敵の面白さ。Fun値が高いほど強い敵である。
     /// </summary>
-    double Fun { get; }
+    int Fun { get; }
     void Initialize(Vector2 spawnPosition, Hero player);
 }
